+++ .github/CHANGELOG.md
@@ -2,12 +2,13 @@
 
 Update <small>_ August 2022</small>
 
+- feat: website command (22/08/2022)
 - feat: add vatsim command (21/08/2022)
 - refactor: remove async from MSFS command and message delete (21/08/2022)
 - refactor: remove async/await from invites (21/08/2022)
 - feat: error handling for process start (21/08/2022)
 - refactor(scamlog): code cleanup and error handling (21/08/2022)
-- docs: improve readme (21/08/2022) 
+- docs: improve readme (21/08/2022)
 - feat: notams command (21/08/2022)
 - refactor: weight and balance command (21/08/2022)
 - refactor: remove async/await from memes (17/08/2022)